Moved the monitoring specific code to a new module: 'MonitoringUtilities', whose functions are used by the CLI and scripts when components are installed/uninstalled
