+++ Core/scripts/dirac-install-db.py
@@ -10,6 +10,7 @@ Create a new DB on the local MySQL server
 __RCSID__ = "$Id$"
 #
 from DIRAC.Core.Utilities import InstallTools
+from DIRAC.FrameworkSystem.Utilities import MonitoringUtilities
 #
 from DIRAC import gConfig
 InstallTools.exitOnError = True
@@ -38,6 +39,6 @@ for db in args:
     InstallTools.addDatabaseOptionsToCS( gConfig, system, db, overwrite = True )
 
     if db != 'InstalledComponentsDB':
-      result = InstallTools.monitorInstallation( 'DB', system, db )
+      result = MonitoringUtilities.monitorInstallation( 'DB', system, db )
       if not result[ 'OK' ]:
         print "ERROR: failed to register installation in database: %s" % result[ 'Message' ]
